Refs #1960. Added micrortps_install.sh script

+++ Tools/generate_microRTPS_bridge.py
@@ -37,7 +37,8 @@ px_generate_uorb_topic_files.append_to_include_path({msg_folder}, px_generate_uo
 out_dir = root_path + "/msgenerated"
 print("Files created in: " + out_dir)
 
-shutil.rmtree(out_dir)
+if os.path.isdir(out_dir):
+    shutil.rmtree(out_dir)
 
 uorb_templates_dir = root_path + "/msg/templates/uorb"
 urtps_templates_dir = root_path + "/msg/templates/urtps"
